+++ react/components/ReviewsContainer.tsx
@@ -16,6 +16,7 @@ const ReviewsContainer: FunctionComponent<ReviewsContainerProps> = ({
   linkText,
   reviews,
   appSettings,
+  relatedProducts,
 }) => {
   return (
     <div className={styles.reviewsContainer}>
@@ -50,7 +51,14 @@ const ReviewsContainer: FunctionComponent<ReviewsContainerProps> = ({
 
       <Fragment>
         {reviews.map((review: any, i: number) => {
-          return <Review review={review} key={i} appSettings={appSettings} />
+          return (
+            <Review
+              review={review}
+              key={i}
+              appSettings={appSettings}
+              relatedProducts={relatedProducts}
+            />
+          )
         })}
       </Fragment>
     </div>
@@ -68,6 +76,7 @@ interface ReviewsContainerProps {
   handleFilter: any
   filter: string
   appSettings: any
+  relatedProducts: any
 }
 
 export default ReviewsContainer