+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Reviews now show if they are from a similar productor if they were originally posted on a third party site
+
 ## [2.1.0] - 2021-03-29
 ### Changed